Refactored: Improved naming. Refactored: Simplified how to add new games. Refactored: The GameImage no longer changes the screen offset, it now just directly controls the images X,Y, this keeps it simpler

+++ src/pixelflut/Core/PixelBuffer.cs
@@ -96,8 +96,8 @@ public class PixelBuffer
         screenProtocol.WriteToBuffer(
             buffer,
             position.position,
-            X + (int)screenConfiguration.OffsetX,
-            Y + (int)screenConfiguration.OffsetY);
+            X + screenConfiguration.OffsetX,
+            Y + screenConfiguration.OffsetY);
     }
 
     public void SetPixel(int pixelNumber, int X, int Y, byte R, byte G, byte B, byte A)
@@ -112,8 +112,8 @@ public class PixelBuffer
         screenProtocol.WriteToBuffer(
             buffer, 
             position.position, 
-            X + (int)screenConfiguration.OffsetX, 
-            Y + (int)screenConfiguration.OffsetY, 
+            X + screenConfiguration.OffsetX, 
+            Y + screenConfiguration.OffsetY, 
             R, 
             G, 
             B,
